+++ esmerald/protocols/logging.py
@@ -1,8 +1,12 @@
-from typing import Any
+from typing import Any, TypeVar
+
+from typing_extensions import Protocol, runtime_checkable
+
+T = TypeVar("T")
 
 
-# Default placeholder logger until properly configured
-class EsmeraldLogger:
+@runtime_checkable
+class LoggingProtocol(Protocol):  # pragma: no cover
     def info(self, *args: Any, **kwargs: Any) -> None: ...
 
     def debug(self, *args: Any, **kwargs: Any) -> None: ...
@@ -12,7 +16,3 @@ class EsmeraldLogger:
     def error(self, *args: Any, **kwargs: Any) -> None: ...
 
     def critical(self, *args: Any, **kwargs: Any) -> None: ...
-
-
-# Global logger used by Esmerald
-logger: EsmeraldLogger = EsmeraldLogger()